+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: oqatim <oqatim@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/12 23:01:12 by oqatim            #+#    #+#             */
-/*   Updated: 2021/11/13 00:08:27 by oqatim           ###   ########.fr       */
+/*   Updated: 2021/11/21 13:43:40 by oqatim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(char *haystack, char *needle, size_t len)
 {
-	int	i;
-	int	o;
+	size_t	i;
+	size_t	o;
 
 	i = 0;
 	while (haystack[i] != '\0')
